# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a crixo cample",
        "title": "Swagger Example API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "cris@swagger.io"
        },
        "license": {},
        "version": "1.0"
    },
    "host": "localhost:8081",
    "paths": {
        "/patients": {
            "get": {
                "description": "Return a paged set of patient profiles w/ the total number of available patients",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "patients"
                ],
                "summary": "Get paged patient profiles",
                "operationId": "patients-profile-find",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.PatientProfile"
                            }
                        },
                        "headers": {
                            "X-Total-Count": {
                                "type": "int",
                                "description": "Total number of records w/o pagination"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Patient Profile description",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "patients"
                ],
                "summary": "Create Patient Profile",
                "operationId": "patients-profile-create",
                "parameters": [
                    {
                        "description": "Create new paatient profile",
                        "name": "patientProfile",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/model.PatientProfile"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/model.PatientProfile"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "model.PatientProfile": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "city": {
                    "type": "string"
                },
                "dateOfBirth": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "firstName": {
                    "type": "string"
                },
                "lastName": {
                    "type": "string"
                },
                "mobile": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "profession": {
                    "type": "string"
                },
                "province": {
                    "type": "string"
                },
                "zipCode": {
                    "type": "string"
                }
            }
        }
    }
}